+++ packages/components/src/Switch/Switch.test.tsx
@@ -64,7 +64,9 @@ test('supports base styling', () => {
   render(<Switch>Label</Switch>);
   const { label, container, track, thumb } = getSwitchParts();
 
-  expect(label.className).toMatchInlineSnapshot(`"flex w-[var(--labelWidth)]"`);
+  expect(label.className).toMatchInlineSnapshot(
+    `"inline-flex w-[var(--labelWidth)]"`
+  );
   expect(container.className).toMatchInlineSnapshot(
     `"w-full group/switch flex items-center gap-[1ch]"`
   );
@@ -102,7 +104,9 @@ test('takes full width by default', () => {
 test('allows to set width via prop', () => {
   render(<Switch width={10}>Label</Switch>);
   const { label } = getSwitchParts();
-  expect(label.className).toMatchInlineSnapshot(`"flex w-[var(--labelWidth)]"`);
+  expect(label.className).toMatchInlineSnapshot(
+    `"inline-flex w-[var(--labelWidth)]"`
+  );
 });
 
 test('supports disabled prop', () => {
@@ -120,20 +124,6 @@ test('renders hidden <input> element', () => {
   render(<Switch>Label</Switch>);
   const { input } = getSwitchParts();
   expect(input instanceof HTMLInputElement).toBeTruthy();
-});
-
-test('toggle switch per click', () => {
-  render(<Switch>Label</Switch>);
-
-  const { input, track } = getSwitchParts();
-
-  fireEvent.click(input);
-  expect(track.className).toMatchInlineSnapshot(`"relative"`);
-  expect(input.checked).toBeTruthy();
-
-  fireEvent.click(input);
-  expect(track.className).toMatchInlineSnapshot(`"relative"`);
-  expect(input.checked).toBeFalsy();
 });
 
 test('focus element and toggle switch per keyboard space', async () => {