+++ themes/theme-core/src/components/SectionMessage.styles.ts
@@ -44,5 +44,5 @@ export const SectionMessage: ThemeComponent<'SectionMessage'> = {
       variant: 'info',
     },
   }),
-  close: cva('h-5 w-5 text-center'),
+  close: cva('size-5 text-center'),
 };