+++ themes/theme-rui/src/components/Pagination.styles.ts
@@ -8,6 +8,7 @@ export const Pagination: ThemeComponent<'Pagination'> = {
     '[&_svg]:pointer-events-none [&_svg]:shrink-0',
     'hover:bg-hover hover:text-hover-foreground',
     'h-9 py-2 gap-1 px-2.5',
+    'cursor-pointer',
   ]),
   pageButton: cva([
     'inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-lg text-sm font-medium transition-colors bg-background size-9',
@@ -16,6 +17,7 @@ export const Pagination: ThemeComponent<'Pagination'> = {
     '[&_svg]:pointer-events-none [&_svg]:shrink-0',
     'data-[selected=true]:border data-[selected=true]:border-input data-[selected=true]:shadow-xs',
     'hover:bg-hover hover:text-hover-foreground',
+    'cursor-pointer',
   ]),
   icon: cva('h-4 w-4'),
 };